requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
future==0.18.2
gcloud==0.18.3
google-api-core==1.24.1
google-auth==1.24.0
google-cloud==0.34.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
jwcrypto==0.8
libcst==0.3.16
line-bot-sdk==1.18.0
lxml==4.6.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.5
oauth2client==4.1.3
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
pandas==1.2.0
Pillow==8.1.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
Pyrebase4==4.4.3
pytesseract==0.3.7
python-dateutil==2.8.1
python-jwt==3.3.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7
scikit-learn==0.24.0
scipy==1.6.0
selenium==3.141.0
six==1.15.0
soupsieve==2.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
XlsxWriter==1.3.7
